feat: Add subdomain extraction and routing for farm-page
- Implemented subdomain extraction from the request hostname to enhance routing capabilities.
- Updated middleware to rewrite URLs for farm-page routes, incorporating the extracted subdomain as a search parameter for improved functionality.
