+++ src/app/components/dashboard/common/suspense.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { Suspense } from "react";
+
+export function WithSuspense({ children }: { children: React.ReactNode }) {
+  return <Suspense fallback={<div>Loading...</div>}>{children}</Suspense>;
+}
