+++ src/app/farmer-dashboard/my-farms/update-farm/page.tsx
@@ -246,8 +246,6 @@ const UpdateFarm = () => {
     }
   };
 
-  console.log(form, uploadedImages);
-
   return (
     <ProtectedRoute requiredUserType="farmer">
       <DashboardLayout>
