+++ src/middleware.ts
@@ -5,6 +5,10 @@ export function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
   const token = request.cookies.get("access")?.value;
 
+  // Extract subdomain from hostname
+  const hostname = request.headers.get("host") || "";
+  const subdomain = hostname.split(".")[0];
+
   // List of protected routes
   const isProtectedRoute =
     pathname.startsWith("/farmer-dashboard") ||
@@ -18,6 +22,13 @@ export function middleware(request: NextRequest) {
     loginUrl.searchParams.set("redirect", pathname);
 
     return NextResponse.redirect(loginUrl);
+  }
+
+  // Add subdomain to search params for farm-page routes
+  if (pathname.startsWith("/farm-page")) {
+    const url = request.nextUrl.clone();
+    url.searchParams.set("farm", subdomain);
+    return NextResponse.rewrite(url);
   }
 
   return NextResponse.next();
